+++ mayan/apps/testing/tests/mixins.py
@@ -589,6 +589,7 @@ class TestModelTestCaseMixin(ContentTypeTestCaseMixin, PermissionTestMixin):
         class Meta:
             app_label = self.app_config.label
             db_table = self._test_db_table
+            ordering = ('id',)
             verbose_name = self._test_model_name
 
         if self.options:
@@ -601,7 +602,7 @@ class TestModelTestCaseMixin(ContentTypeTestCaseMixin, PermissionTestMixin):
         def save(instance, *args, **kwargs):
             # Custom .save() method to use random primary key values.
             if instance.pk:
-                return models.Model.self(instance, *args, **kwargs)
+                return models.Model.save(instance, *args, **kwargs)
             else:
                 instance.pk = RandomPrimaryKeyModelMonkeyPatchMixin.get_unique_primary_key(
                     model=instance._meta.model
@@ -641,7 +642,6 @@ class TestServerTestCaseMixin:
 
 class TestViewTestCaseMixin:
     auto_add_test_view = False
-    has_test_view = False
     test_view_is_public = False
     test_view_object = None
     test_view_name = TEST_VIEW_NAME
@@ -650,15 +650,18 @@ class TestViewTestCaseMixin:
 
     def setUp(self):
         super().setUp()
+        self._test_view_count = 0
+        self._test_view_names = []
+
         if self.auto_add_test_view:
             self.add_test_view(test_object=self.test_view_object)
 
     def tearDown(self):
-        urlconf = importlib.import_module(settings.ROOT_URLCONF)
-
         self.client.logout()
-        if self.has_test_view:
-            urlconf.urlpatterns.pop(0)
+
+        for _ in range(self._test_view_count):
+            self._get_test_view_urlpatterns().pop(0)
+
         super().tearDown()
 
     def _get_context_from_test_response(self, response):
@@ -690,19 +693,37 @@ class TestViewTestCaseMixin:
         else:
             return test_view
 
-    def add_test_view(self, test_object=None):
-        urlconf = importlib.import_module(settings.ROOT_URLCONF)
+    def _get_test_view_urlpatterns(self):
+        return importlib.import_module(settings.ROOT_URLCONF).urlpatterns
 
-        urlconf.urlpatterns.insert(
+    def add_test_view(
+        self, test_object=None, test_view_factory=None, test_view_name=None,
+        test_view_url=None
+    ):
+        if test_view_factory:
+            view = test_view_factory()
+        else:
+            view = self._test_view_factory(
+                test_object=test_object
+            )
+
+        if test_view_name:
+            self._test_view_name = test_view_name
+        else:
+            self._test_view_name = '{}_{}'.format(
+                TEST_VIEW_NAME, len(self._test_view_names)
+            )
+
+        self._get_test_view_urlpatterns().insert(
             0, url(
-                regex=self.test_view_url, view=self._test_view_factory(
-                    test_object=test_object
-                ), name=self.test_view_name
+                regex=test_view_url or self.test_view_url, view=view,
+                name=self._test_view_name
             )
         )
         clear_url_caches()
-        self.has_test_view = True
+        self._test_view_count += 1
+        self._test_view_names.append(self._test_view_name)
 
     def get_test_view(self):
-        response = self.get(viewname=self.test_view_name)
+        response = self.get(viewname=self._test_view_name)
         return self._get_context_from_test_response(response=response)