+++ mayan/apps/cabinets/apps.py
@@ -2,12 +2,13 @@ from django.apps import apps
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.classes import ModelPermission
-from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
+from mayan.apps.acls.permissions import (
+    permission_acl_edit, permission_acl_view
+)
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.classes import ModelCopy, ModelQueryFields
 from mayan.apps.common.menus import (
-    menu_facet, menu_list_facet, menu_main, menu_multi_item, menu_object,
-    menu_secondary
+    menu_list_facet, menu_main, menu_multi_item, menu_object, menu_secondary
 )
 from mayan.apps.events.classes import EventModelRegistry, ModelEventType
 from mayan.apps.navigation.classes import SourceColumn
@@ -102,6 +103,11 @@ class CabinetsApp(MayanAppConfig):
                 event_cabinet_document_removed
             )
         )
+        ModelEventType.register(
+            model=Document, event_types=(
+                event_cabinet_document_added, event_cabinet_document_removed
+            )
+        )
 
         ModelPermission.register(
             model=Document, permissions=(
@@ -139,10 +145,6 @@ class CabinetsApp(MayanAppConfig):
         )
 
         SourceColumn(
-            attribute='label', is_identifier=True, is_sortable=True,
-            source=CabinetSearchResult
-        )
-        SourceColumn(
             attribute='get_full_path', source=CabinetSearchResult
         )
 
@@ -168,7 +170,7 @@ class CabinetsApp(MayanAppConfig):
             widget=DocumentCabinetWidget
         )
 
-        menu_facet.bind_links(
+        menu_list_facet.bind_links(
             links=(link_document_cabinet_list,), sources=(Document,)
         )