+++ mayan/apps/document_states/apps.py
@@ -8,8 +8,8 @@ from mayan.apps.common.classes import (
     ModelCopy, ModelField, ModelProperty, ModelReverseField
 )
 from mayan.apps.common.menus import (
-    menu_facet, menu_list_facet, menu_main, menu_multi_item, menu_object,
-    menu_related, menu_secondary, menu_setup, menu_tools
+    menu_list_facet, menu_main, menu_multi_item, menu_object, menu_related,
+    menu_secondary, menu_setup, menu_tools
 )
 from mayan.apps.documents.links.document_type_links import link_document_type_list
 from mayan.apps.documents.signals import signal_post_document_type_change
@@ -435,7 +435,7 @@ class DocumentStatesApp(MayanAppConfig):
             source=WorkflowStateRuntimeProxy
         )
 
-        menu_facet.bind_links(
+        menu_list_facet.bind_links(
             links=(link_workflow_instance_list,), sources=(Document,)
         )
         menu_secondary.bind_links(