+++ mayan/apps/metadata/apps.py
@@ -5,13 +5,15 @@ from django.db.models.signals import post_delete, post_save
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.classes import ModelPermission
-from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
+from mayan.apps.acls.permissions import (
+    permission_acl_edit, permission_acl_view
+)
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.classes import (
     ModelCopy, ModelFieldRelated, ModelProperty, ModelQueryFields
 )
 from mayan.apps.common.menus import (
-    menu_facet, menu_list_facet, menu_multi_item, menu_object, menu_related,
+    menu_list_facet, menu_multi_item, menu_object, menu_related,
     menu_secondary, menu_setup
 )
 from mayan.apps.documents.links.document_type_links import link_document_type_list
@@ -237,7 +239,7 @@ class MetadataApp(MayanAppConfig):
 
         # Document metadata
 
-        menu_facet.bind_links(
+        menu_list_facet.bind_links(
             links=(link_metadata_view,), sources=(Document,)
         )
         menu_multi_item.bind_links(