+++ mayan/apps/navigation/classes.py
@@ -467,6 +467,7 @@ class Menu(TemplateObjectMixin):
 
     def resolve(self, context=None, request=None, source=None, sort_results=False):
         result = []
+        self.matched_link_set = set()
 
         if not context and not request:
             raise ImproperlyConfigured(
@@ -560,6 +561,12 @@ class Menu(TemplateObjectMixin):
         result = []
 
         object_resolved_links = []
+
+        # Deduplicate matched links.
+        matched_links = list(
+            set(matched_links).difference(self.matched_link_set)
+        )
+        self.matched_link_set.update(matched_links)
 
         for link in matched_links:
             kwargs = {