+++ mayan/apps/rest_api/serializers.py
@@ -2,6 +2,17 @@ from rest_framework import serializers
 from rest_framework.reverse import reverse
 
 
+class BatchAPIRequestResponseSerializer(serializers.Serializer):
+    data = serializers.JSONField(read_only=True)
+    headers = serializers.DictField(read_only=True)
+    name = serializers.CharField(read_only=True)
+    status_code = serializers.IntegerField(read_only=True)
+    requests = serializers.JSONField(
+        style={'base_template': 'textarea.html'},
+        write_only=True
+    )
+
+
 class BlankSerializer(serializers.Serializer):
     """Serializer for the object action API view."""