+++ mayan/apps/user_management/methods.py
@@ -106,6 +106,12 @@ def method_user_get_absolute_url(self):
     )
 
 
+def method_user_get_absolute_api_url(self):
+    return reverse(
+        viewname='rest_api:user-detail', kwargs={'user_id': self.pk}
+    )
+
+
 def method_user_get_groups(self, user, permission=permission_group_view):
     AccessControlList = apps.get_model(
         app_label='acls', model_name='AccessControlList'