+++ mayan/apps/web_links/apps.py
@@ -2,12 +2,13 @@ from django.apps import apps
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.classes import ModelPermission
-from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
+from mayan.apps.acls.permissions import (
+    permission_acl_edit, permission_acl_view
+)
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.classes import ModelCopy
 from mayan.apps.common.menus import (
-    menu_facet, menu_list_facet, menu_object, menu_related, menu_secondary,
-    menu_setup
+    menu_list_facet, menu_object, menu_related, menu_secondary, menu_setup
 )
 from mayan.apps.documents.links.document_type_links import (
     link_document_type_list
@@ -65,6 +66,11 @@ class WebLinksApp(MayanAppConfig):
 
         ModelEventType.register(
             event_types=(
+                event_web_link_navigated,
+            ), model=Document
+        )
+        ModelEventType.register(
+            event_types=(
                 event_web_link_edited, event_web_link_navigated
             ), model=WebLink
         )
@@ -102,7 +108,7 @@ class WebLinksApp(MayanAppConfig):
         )
         source_column_enabled.add_exclude(source=ResolvedWebLink)
 
-        menu_facet.bind_links(
+        menu_list_facet.bind_links(
             links=(link_document_web_link_list,),
             sources=(Document,)
         )