+++ HISTORY.rst
@@ -354,6 +354,53 @@
   milliseconds. Use `None` to disable.
 - Enforce ``CONVERTER_IMAGE_GENERATION_MAX_RETRIES`` setting and add logging
   message when the maximum retires are exhausted.
+- Messaging app updates:
+
+  - Add API views.
+  - Exclude superusers and staff users from being message recipients.
+  - Add dedicated create message form.
+  - Use Select2 for the user selection field.
+  - Add message edit permission. This permission is required in order to
+    change the message read status.
+
+- Add ``get_absolute_api_url`` method to download files, document versions
+  and users. These URL are used to determine the message sender's API URL.
+- Test view mixin updates:
+
+  - Add a default ordering to the ``TestModel`` to silence warning.
+  - Fix ``TestModel.save()`` method.
+  - Support multiple test views per test case.
+  - Allow subclasses to supply their own ``urlpatterns``.
+  - Support passing arguments to ``add_test_view``.
+
+- Add batch API request support.
+- Adjust event registrations:
+
+  - Register cabinet document add and remove events to the Document model too.
+  - Register document file parsing events to the Document model too.
+  - Rename label of the document parsed content deleted event.
+  - Replace the ``DownloadFile`` content object registration from the
+    ``Document`` model to the ``DocumentVersion`` model.
+  - Register the document file created, edited events to the ``Document``
+    model too.
+  - Register the document version created, edited events to the ``Document``
+    model too.
+  - Register the document trashed event to the ``Document`` model too.
+  - Remove the document file created event from the ``DocumentFile`` model.
+  - Remove the document version created event from the ``DocumentVersion``
+    model.
+  - Register the document version page deleted event to the
+    ``DocumentVersion`` model.
+  - Remove the document version page deleted event from the
+    ``DocumentVersionPage`` model.
+  - Register the tag attached, removed events to the ``Document`` model too.
+  - Register the web link navigated event to the ``Document`` model too.
+  - Remove the document version page OCR edited event from the ``Document``
+    model.
+  - Register the document version OCR submitted, finished, content deleted
+    events to the ``DocumentVersion`` model.
+
+- Sort object and list facet links when using the list item view.
 
 4.0.16 (2021-09-29)
 ===================